refacto: split components and pages
