+++ src/components/PlayerProfil/PlayerProfil.tsx
@@ -21,7 +21,7 @@ import {
 import { BrowserView, MobileView } from "react-device-detect";
 import { useEffect, useState } from "react";
 import { useAppDispatch, useAppSelector } from "../../hooks/redux";
-import { getPlayerInfos } from "../../Redux-store/Redux-reducers/player";
+import { getPlayerInfos } from "../../redux/Redux-reducers/player";
 import crud from "@/utils/crud";
 
 import "./PlayerProfil.scss";
