+++ src/pages/Scout/Scout.scss
