+++ src/pages/Scout/SearchPlayer.scss
