+++ src/redux/Redux-actions/player.tsx
