+++ src/redux/Redux-reducers/player.tsx
