+++ src/redux/Redux-reducers/scout.tsx
