+++ src/redux/index.tsx
